Refactor rounds tests into own files in participant demo (#20)
