+++ participant/src/tests.rs
@@ -1,2 +1,4 @@
 mod cli;
 mod lib;
+mod round1;
+mod round2;
